refactor: extract chat message display helpers + restoreVisibleSwipeState into useChatMessageDisplay
