+++ src/views/ChatView.vue
@@ -95,6 +95,7 @@ import { showToast } from '@/core/states/toastState.js';
 import { triggerAutoSyncCheck } from '@/composables/chat/useAutoSync.js';
 import { useMemoryBooks } from '@/composables/chat/useMemoryBooks.js';
 import { useMemoryAutomation } from '@/composables/chat/useMemoryAutomation.js';
+import { useChatMessageDisplay, restoreVisibleSwipeState } from '@/composables/chat/useChatMessageDisplay.js';
 import { getMemoryPromptOptions, getMemoryPromptLabel, getMemoryPromptLabelByKey, getNormalizedMemoryGenerationState } from '@/core/services/memoryPromptPresets.js';
 import * as memoryBooksService from '@/core/services/memoryBooksService.js';
 import * as contextService from '@/core/services/contextService.js';
@@ -343,6 +344,14 @@ const {
     openMemoryBooksSheet
 });
 
+const {
+    getAvatar,
+    getAvatarLetter,
+    getAvatarColor,
+    getDisplayName,
+    openAvatar
+} = useChatMessageDisplay(activeChatChar, allPersonas);
+
 const onRegexChanged = () => { regexRevision.value++; };
 const contextBreakdown = ref(null);
 // Import context settings from service
@@ -381,31 +390,6 @@ const selectionIncludesLast = computed(() => {
 });
 
 
-function restoreVisibleSwipeState(messages = []) {
-    if (!Array.isArray(messages)) return [];
-
-    return messages.map(msg => {
-        if (!msg || !Array.isArray(msg.swipesMeta)) return msg;
-
-        const swipeIndex = msg.swipeId || 0;
-        const swipeMeta = msg.swipesMeta[swipeIndex];
-        if (!swipeMeta) return msg;
-
-        let nextMsg = msg;
-
-        if (msg.reasoning == null && swipeMeta.reasoning != null) {
-            nextMsg = { ...nextMsg, reasoning: swipeMeta.reasoning };
-        }
-        if (nextMsg.genTime == null && swipeMeta.genTime != null) {
-            nextMsg = { ...nextMsg, genTime: swipeMeta.genTime };
-        }
-        if ((nextMsg.tokens == null || nextMsg.tokens === 0) && swipeMeta.tokens != null) {
-            nextMsg = { ...nextMsg, tokens: swipeMeta.tokens };
-        }
-
-        return nextMsg;
-    });
-}
 
 watch([isSearchMode, isSelectionMode], () => {
     ignoreScrollAdjustment = true;
@@ -3174,45 +3158,6 @@ async function startImpersonation(guidanceText = null) {
 }
 
 // --- Utils ---
-
-function getAvatar(msg) {
-    if (msg.role === 'user') {
-        // Try to resolve avatar from allPersonas by ID (IndexedDB-backed)
-        if (msg.persona?.id) {
-            const p = allPersonas.value.find(p => p.id === msg.persona.id);
-            if (p?.avatar) return p.avatar;
-        }
-        // Fallback to embedded avatar for old messages
-        return msg.persona?.avatar || null;
-    }
-    return activeChatChar?.avatar || null;
-}
-
-function getAvatarLetter(msg) {
-    if (msg.role === 'user') return (msg.persona?.name?.[0] || "U").toUpperCase();
-    return (activeChatChar?.name?.[0] || "?").toUpperCase();
-}
-
-function getAvatarColor(msg) {
-    if (msg.role === 'user') return 'var(--vk-blue)';
-    return activeChatChar?.color || '#ccc';
-}
-
-function getDisplayName(msg) {
-    if (msg.role === 'user') return msg.persona?.name || "User";
-    return activeChatChar?.name || "Character";
-}
-
-function openAvatar(msg) {
-    const src = getAvatar(msg);
-    if (src) {
-        const name = getDisplayName(msg);
-        const description = "";
-        window.dispatchEvent(new CustomEvent('trigger-open-image', { 
-            detail: { src, name, description, onCloseCallback: null } 
-        }));
-    }
-}
 
 async function deleteSession(sessionId, targetChar) {
     const char = targetChar || activeChatChar;
